+++ src/service/impl/quays/jp3/quays-coordinates.graphql
